- Split BackgroundGL out as a widget of its own - wf-background writes to user temporary cache when background image changes - locker picks up background image changes and matches - recrustify - wf-background inhibit output is now a user option
