+++ metadata/background.xml.in
@@ -21,6 +21,10 @@
 		<_short>Randomize</_short>
 		<default>false</default>
 	</option>
+	<option name="blank" type="bool">
+		<_short>Blank screen on startup</_short>
+		<default>true</default>
+	</option>
 	<option name="fill_mode" type="string">
 		<_short>Fill mode</_short>
 		<default>stretch</default>
