+++ metadata/locker.xml
@@ -5,6 +5,10 @@
 	<category>Shell</category>
 	<group>
 	<_short>General</_short>
+	<option name="background_image" type="bool">
+		<_short>Use wf-background image</_short>
+		<default>true</default>
+	</option>
 	<option name="background_color" type="string">
 		<_short>Background Color</_short>
 		<default>gtk_headerbar</default>
